dev: trie - rename util.ts to trie-util.ts
