+++ packages/cspell-trie-lib/src/lib/importExportV1.ts
@@ -2,7 +2,7 @@ import { TrieNode, FLAG_WORD, ChildMap, TrieRoot } from './TrieNode';
 import { TrieRefNode } from './trieRef';
 import { Sequence, genSequence } from 'gensequence';
 import { convertToTrieRefNodes } from './convertToTrieRefNodes';
-import { trieNodeToRoot } from './util';
+import { trieNodeToRoot } from './trie-util';
 
 const EOW = '*';
 export const DATA = EOW;
