+++ packages/cspell-trie-lib/src/lib/index.ts
@@ -1,6 +1,6 @@
 export * from './trie';
 export { TrieNode, FLAG_WORD, ChildMap, TrieRoot } from './TrieNode';
-export * from './util';
+export * from './trie-util';
 export * from './walker';
 export * from './importExport';
 export { buildTrie, buildTrieFast, TrieBuilder } from './TrieBuilder';
