+++ packages/cspell-trie-lib/src/lib/trie-util.ts
