+++ packages/cspell-trie2-lib/README.md
@@ -1,3 +1,5 @@
 # cspell-trie2
 
 Trie library for cspell
+
+This package has be deprecated.
