+++ packages/cspell-trie2-lib/package.json
@@ -2,6 +2,7 @@
   "name": "cspell-trie2-lib",
   "version": "5.15.2",
   "description": "Trie Data Structure to support cspell.",
+  "private": true,
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
